@@ -114,7 +114,6 @@ func validatePackages(config configuration.Config) (err error) {
 	defer timestamp.StopEvent(nil)
 
 	const (
-		selinuxPkgName     = "selinux-policy"
 		validateError      = "failed to validate package lists in config"
 		verityPkgName      = "verity-read-only-root"
 		verityDebugPkgName = "verity-read-only-root-debug-tools"
@@ -132,6 +131,11 @@ func validatePackages(config configuration.Config) (err error) {
 		foundVerityInitramfsDebugPackage := false
 		foundDracutFipsPackage := false
 		kernelCmdLineString := systemConfig.KernelCommandLine.ExtraCommandLine
+		selinuxPkgName := systemConfig.KernelCommandLine.SELinuxPolicy
+		if selinuxPkgName == "" {
+			selinuxPkgName = "selinux-policy"
+		}
+
 		for _, pkg := range packageList {
 			if pkg == "kernel" {
 				return fmt.Errorf("%s: kernel should not be included in a package list, add via config file's [KernelOptions] entry", validateError)
