@@ -8,6 +8,7 @@ package diskutils
 import (
 	"encoding/json"
 	"fmt"
+	"os"
 	"path/filepath"
 	"regexp"
 	"strconv"
@@ -39,9 +40,36 @@ type SystemBlockDevice struct {
 	Model       string // Example: Virtual Disk
 }
 
+type partitionInfoOutput struct {
+	Devices []PartitionInfo `json:"blockdevices"`
+}
+
+type PartitionInfo struct {
+	Name              string `json:"name"`       // Example: nbd0p1
+	Path              string `json:"path"`       // Example: /dev/nbd0p1
+	PartitionTypeUuid string `json:"parttype"`   // Example: c12a7328-f81f-11d2-ba4b-00a0c93ec93b
+	FileSystemType    string `json:"fstype"`     // Example: vfat
+	Uuid              string `json:"uuid"`       // Example: 4BD9-3A78
+	PartUuid          string `json:"partuuid"`   // Example: 7b1367a6-5845-43f2-99b1-a742d873f590
+	Mountpoint        string `json:"mountpoint"` // Example: /mnt/os/boot
+	PartLabel         string `json:"partlabel"`  // Example: boot
+}
+
+type loopbackListOutput struct {
+	Devices []loopbackDevice `json:"loopdevices"`
+}
+
+type loopbackDevice struct {
+	Name        string `json:"name"`
+	BackingFile string `json:"back-file"`
+}
+
 const (
 	// AutoEndSize is used as the disk's "End" value to indicate it should be picked automatically
 	AutoEndSize = 0
+
+	EfiSystemPartitionTypeUuid = "c12a7328-f81f-11d2-ba4b-00a0c93ec93b"
+	BiosBootPartitionTypeUuid  = "21686148-6449-6e6f-744e-656564454649"
 )
 
 const (
@@ -188,37 +216,32 @@ func ApplyRawBinary(diskDevPath string, rawBinary configuration.RawBinary) (err 
 }
 
 // CreateEmptyDisk creates an empty raw disk in the given working directory as described in disk configuration
-func CreateEmptyDisk(workDirPath, diskName string, disk configuration.Disk) (diskFilePath string, err error) {
-	const (
-		defautBlockSize = MiB
-	)
+func CreateEmptyDisk(workDirPath, diskName string, maxSize uint64) (diskFilePath string, err error) {
 	diskFilePath = filepath.Join(workDirPath, diskName)
 
-	// Assume that Disk.MaxSize is given
-	maxSize := disk.MaxSize
-	err = sparseDisk(diskFilePath, defautBlockSize, maxSize)
+	err = CreateSparseDisk(diskFilePath, maxSize, 0o644)
 	return
 }
 
-// sparseDisk creates an empty sparse disk file.
-func sparseDisk(diskPath string, blockSize, size uint64) (err error) {
-	ddArgs := []string{
-		"if=/dev/zero",                  // Input file.
-		fmt.Sprintf("of=%s", diskPath),  // Output file.
-		fmt.Sprintf("bs=%d", blockSize), // Size of one copied block.
-		fmt.Sprintf("seek=%d", size),    // Size of the image.
-		"count=0",                       // Number of blocks to copy to the output file.
+// CreateSparseDisk creates an empty sparse disk file.
+func CreateSparseDisk(diskPath string, size uint64, perm os.FileMode) (err error) {
+	// Open and truncate the file.
+	file, err := os.OpenFile(diskPath, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, perm)
+	if err != nil {
+		return fmt.Errorf("failed to create empty disk file:\n%w", err)
 	}
 
-	_, stderr, err := shell.Execute("dd", ddArgs...)
+	// Resize the file to the desired size.
+	err = file.Truncate(int64(size * MiB))
 	if err != nil {
-		logger.Log.Warnf("Failed to create empty disk with dd: %v", stderr)
+		return fmt.Errorf("failed to set empty disk file's size:\n%w", err)
 	}
 	return
 }
 
 // SetupLoopbackDevice creates a /dev/loop device for the given disk file
 func SetupLoopbackDevice(diskFilePath string) (devicePath string, err error) {
+	logger.Log.Debugf("Attaching Loopback: %v", diskFilePath)
 	stdout, stderr, err := shell.Execute("losetup", "--show", "-f", "-P", diskFilePath)
 	if err != nil {
 		logger.Log.Warnf("Failed to create loopback device using losetup: %v", stderr)
@@ -231,44 +254,61 @@ func SetupLoopbackDevice(diskFilePath string) (devicePath string, err error) {
 
 // BlockOnDiskIO waits until all outstanding operations against a disk complete.
 func BlockOnDiskIO(diskDevPath string) (err error) {
-	const (
-		// Indices for values in /proc/diskstats
-		majIdx            = 0
-		minIdx            = 1
-		outstandingOpsIdx = 11
-	)
-	var blockDevices blockDevicesOutput
-
-	logger.Log.Infof("Flushing all IO to disk for %s", diskDevPath)
-	_, _, err = shell.Execute("sync")
+	maj, min, err := GetDiskIds(diskDevPath)
 	if err != nil {
 		return
 	}
 
+	return BlockOnDiskIOByIds(diskDevPath, maj, min)
+}
+
+func GetDiskIds(diskDevPath string) (maj string, min string, err error) {
 	rawDiskOutput, stderr, err := shell.Execute("lsblk", "--nodeps", "--json", "--output", "NAME,MAJ:MIN", diskDevPath)
 	if err != nil {
 		logger.Log.Warn(stderr)
+		err = fmt.Errorf("failed to find IDs for disk (%s):\n%w", diskDevPath, err)
 		return
 	}
 
 	bytes := []byte(rawDiskOutput)
+
+	var blockDevices blockDevicesOutput
 	err = json.Unmarshal(bytes, &blockDevices)
 	if err != nil {
 		return
 	}
 
 	if len(blockDevices.Devices) != 1 {
-		return fmt.Errorf("couldn't find disk IDs for %s (%s), expecting only one result", diskDevPath, rawDiskOutput)
+		err = fmt.Errorf("couldn't find disk IDs for %s (%s), expecting only one result", diskDevPath, rawDiskOutput)
+		return
 	}
 	// MAJ:MIN is returned in the form "1:2"
 	diskIDs := strings.Split(blockDevices.Devices[0].MajMin, ":")
 	if len(diskIDs) != 2 {
-		return fmt.Errorf("couldn't find disk IDs for %s (%s), couldn't parse MAJ:MIN", diskDevPath, rawDiskOutput)
+		err = fmt.Errorf("couldn't find disk IDs for %s (%s), couldn't parse MAJ:MIN", diskDevPath, rawDiskOutput)
+		return
 	}
-	maj := diskIDs[0]
-	min := diskIDs[1]
+	maj = diskIDs[0]
+	min = diskIDs[1]
+	return
+}
 
-	logger.Log.Tracef("Searching /proc/diskstats for %s (%s:%s)", blockDevices.Devices[0].Name, maj, min)
+// BlockOnDiskIOById waits until all outstanding operations against a disk complete.
+func BlockOnDiskIOByIds(debugName string, maj string, min string) (err error) {
+	const (
+		// Indices for values in /proc/diskstats
+		majIdx            = 0
+		minIdx            = 1
+		outstandingOpsIdx = 11
+	)
+
+	logger.Log.Debugf("Flushing all IO to disk")
+	_, _, err = shell.Execute("sync")
+	if err != nil {
+		return
+	}
+
+	logger.Log.Tracef("Searching /proc/diskstats for %s (%s:%s)", debugName, maj, min)
 	for {
 		var (
 			foundEntry     = false
@@ -293,13 +333,12 @@ func BlockOnDiskIO(diskDevPath string) (err error) {
 
 		err = shell.ExecuteLiveWithCallback(onStdout, logger.Log.Error, true, "cat", "/proc/diskstats")
 		if err != nil {
-			logger.Log.Error(stderr)
 			return
 		}
 		if !foundEntry {
-			return fmt.Errorf("couldn't find entry for '%s' in /proc/diskstats", diskDevPath)
+			return fmt.Errorf("couldn't find entry for '%s' in /proc/diskstats", debugName)
 		}
-		logger.Log.Debugf("Outstanding operations on '%s': %s", diskDevPath, outstandingOps)
+		logger.Log.Debugf("Outstanding operations on '%s': %s", debugName, outstandingOps)
 
 		if outstandingOps == "0" {
 			break
@@ -313,12 +352,61 @@ func BlockOnDiskIO(diskDevPath string) (err error) {
 
 // DetachLoopbackDevice detaches the specified disk
 func DetachLoopbackDevice(diskDevPath string) (err error) {
-	logger.Log.Infof("Detaching Loopback Device Path: %v", diskDevPath)
+	logger.Log.Debugf("Detaching Loopback Device Path: %v", diskDevPath)
 	_, stderr, err := shell.Execute("losetup", "-d", diskDevPath)
 	if err != nil {
 		logger.Log.Warnf("Failed to detach loopback device using losetup: %v", stderr)
 	}
 	return
+}
+
+func WaitForLoopbackToDetach(devicePath string, diskPath string) error {
+	if !filepath.IsAbs(diskPath) {
+		return fmt.Errorf("internal error: loopback disk path must be absolute (%s)", diskPath)
+	}
+
+	delay := 100 * time.Millisecond
+	attempts := 5
+	for failures := 0; failures < attempts; failures++ {
+		stdout, _, err := shell.Execute("losetup", "--list", "--json", "--output", "NAME,BACK-FILE")
+		if err != nil {
+			return fmt.Errorf("failed to read loopback list:\n%w", err)
+		}
+
+		var output loopbackListOutput
+		err = json.Unmarshal([]byte(stdout), &output)
+		if err != nil {
+			return fmt.Errorf("failed to parse loopback devices list JSON:\n%w", err)
+		}
+
+		found := false
+		for _, device := range output.Devices {
+			if device.Name == devicePath && device.BackingFile == diskPath {
+				found = true
+				break
+			}
+		}
+
+		if !found {
+			return nil
+		}
+
+		time.Sleep(delay)
+		delay *= 2
+	}
+
+	return fmt.Errorf("timed out waiting for loopback device (%s) for disk (%s) to close", devicePath, diskPath)
+}
+
+// WaitForDevicesToSettle waits for all udev events to be processed on the system.
+// This can be used to wait for partitions to be discovered after mounting a disk.
+func WaitForDevicesToSettle() error {
+	logger.Log.Debugf("Waiting for devices to settle")
+	_, _, err := shell.Execute("udevadm", "settle")
+	if err != nil {
+		return fmt.Errorf("failed to wait for devices to settle:\n%w", err)
+	}
+	return nil
 }
 
 // CreatePartitions creates partitions on the specified disk according to the disk config
@@ -661,6 +749,28 @@ func SystemBlockDevices() (systemDevices []SystemBlockDevice, err error) {
 	}
 
 	return
+}
+
+func GetDiskPartitions(diskDevPath string) ([]PartitionInfo, error) {
+	// Just in case the disk was only recently connected, wait for the OS to finish processing it.
+	err := WaitForDevicesToSettle()
+	if err != nil {
+		return nil, fmt.Errorf("failed to list disk (%s) partitions:\n%w", diskDevPath, err)
+	}
+
+	// Read the disk's partitions.
+	jsonString, _, err := shell.Execute("lsblk", diskDevPath, "--output", "NAME,PATH,PARTTYPE,FSTYPE,UUID,MOUNTPOINT,PARTUUID,PARTLABEL", "--json", "--list")
+	if err != nil {
+		return nil, fmt.Errorf("failed to list disk (%s) partitions:\n%w", diskDevPath, err)
+	}
+
+	var output partitionInfoOutput
+	err = json.Unmarshal([]byte(jsonString), &output)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse disk (%s) partitions JSON:\n%w", diskDevPath, err)
+	}
+
+	return output.Devices, err
 }
 
 func createExtendedPartition(diskDevPath string, partitionTableType string, partitions []configuration.Partition, partIDToFsTypeMap, partDevPathMap map[string]string) (err error) {
