@@ -156,7 +156,17 @@ func cloneSystemConfigs(cloner repocloner.RepoCloner, configFile, baseDirPath st
 
 	logger.Log.Infof("Cloning: %v", packageVersionsInConfig)
 	// The image tools don't care if a package was created locally or not, just that it exists. Disregard if it is prebuilt or not.
-	_, err = cloner.Clone(cloneDeps, packageVersionsInConfig...)
+	_, err = cloner.CloneByPackageVerSingleTransaction(cloneDeps, packageVersionsInConfig...)
+	if err != nil {
+		// Fallback to legacy flow with multiple transactions in case we get a OOM error from a large transaction.
+		logger.Log.Warnf("Failed to clone packages in a single transaction, will retry with individual transactions... (%s)", err)
+		logger.Log.Warnf("\tCheck log file '%s' for more details from package manager.", *logFile)
+		_, err = cloner.CloneByPackageVer(cloneDeps, packageVersionsInConfig...)
+		if err != nil {
+			logger.Log.Errorf("Also failed to clone packages with individual transactions. Error: %s", err)
+			logger.Log.Errorf("\tCheck log file '%s' for more details from package manager.", *logFile)
+		}
+	}
 	return
 }
 
