@@ -38,10 +38,10 @@ func CurrentEnvironment() []string {
 	return currentEnv
 }
 
-// PermanentlyStopAllProcesses will send the provided signal to all processes spawned by this package,
+// PermanentlyStopAllChildProcesses will send the provided signal to all processes spawned by this package,
 // and all of those process's children.
 // Invoking this will also block future process creation, causing the Execute methods to return an error.
-func PermanentlyStopAllProcesses(signal unix.Signal) {
+func PermanentlyStopAllChildProcesses(signal unix.Signal) {
 	// Acquire the global activeCommandsMutex to ensure no
 	// new commands are executed during this teardown routine
 	logger.Log.Info("Waiting for outstanding processes to be created")
@@ -72,6 +72,11 @@ func PermanentlyStopAllProcesses(signal unix.Signal) {
 
 // Execute runs the provided command.
 func Execute(program string, args ...string) (stdout, stderr string, err error) {
+	return ExecuteInDirectory("", program, args...)
+}
+
+// Execute runs the provided command in a specific working directory.
+func ExecuteInDirectory(workingDirectory, program string, args ...string) (stdout, stderr string, err error) {
 	var (
 		outBuf bytes.Buffer
 		errBuf bytes.Buffer
@@ -80,6 +85,10 @@ func Execute(program string, args ...string) (stdout, stderr string, err error) 
 	cmd := exec.Command(program, args...)
 	cmd.Stdout = &outBuf
 	cmd.Stderr = &errBuf
+
+	if workingDirectory != "" {
+		cmd.Dir = workingDirectory
+	}
 
 	err = trackAndStartProcess(cmd)
 	if err != nil {
@@ -132,6 +141,30 @@ func ExecuteLive(squashErrors bool, program string, args ...string) (err error) 
 	return ExecuteLiveWithCallback(onStdout, onStderr, false, program, args...)
 }
 
+// ExecuteLiveWithErr runs a command in the shell and logs it in real-time.
+// In addition, if there is an error, the last x lines of stderr will be attached to the err object.
+func ExecuteLiveWithErr(stderrLines int, program string, args ...string) (err error) {
+	stderrChan := make(chan string, stderrLines)
+
+	err = ExecuteLiveWithCallbackAndChannels(logger.Log.Debug, logger.Log.Debug, nil, stderrChan, program, args...)
+	close(stderrChan)
+	if err != nil {
+		errLines := ""
+		for errLine := range stderrChan {
+			if errLines != "" {
+				errLines += "\n"
+			}
+			errLines += errLine
+		}
+
+		if errLines != "" {
+			err = fmt.Errorf("%s\n%w", errLines, err)
+		}
+		return
+	}
+	return nil
+}
+
 // ExecuteLiveWithCallback runs a command in the shell and invokes the provided callbacks in real-time on each line of stdout and stderr.
 // If printOutputOnError is true, the full output of the command will be printed after completion if the command returns an error. In the event
 // the buffer becomes full the oldest buffered output is discarded.
@@ -139,6 +172,33 @@ func ExecuteLiveWithCallback(onStdout, onStderr func(...interface{}), printOutpu
 	var outputChan chan string
 	const outputChanBufferSize = 1500
 
+	if printOutputOnError {
+		outputChan = make(chan string, outputChanBufferSize)
+	}
+
+	err = ExecuteLiveWithCallbackAndChannels(onStdout, onStderr, outputChan, outputChan, program, args...)
+	if err != nil {
+		return
+	}
+
+	// Optionally dump the output in the event of an error
+	if outputChan != nil {
+		close(outputChan)
+	}
+	if err != nil && printOutputOnError {
+		logger.Log.Errorf("Call to %s returned error, last %d lines of output:", program, outputChanBufferSize)
+		for line := range outputChan {
+			logger.Log.Warn(line)
+		}
+	}
+
+	return
+}
+
+func ExecuteLiveWithCallbackAndChannels(onStdout, onStderr func(...interface{}),
+	stdoutChannel, stderrChannel chan string,
+	program string, args ...string,
+) (err error) {
 	cmd := exec.Command(program, args...)
 
 	stdoutPipe, err := cmd.StdoutPipe()
@@ -165,25 +225,11 @@ func ExecuteLiveWithCallback(onStdout, onStderr func(...interface{}), printOutpu
 	wg := new(sync.WaitGroup)
 	wg.Add(2)
 
-	if printOutputOnError {
-		outputChan = make(chan string, outputChanBufferSize)
-	}
-	go logger.StreamOutput(stdoutPipe, onStdout, wg, outputChan)
-	go logger.StreamOutput(stderrPipe, onStderr, wg, outputChan)
+	go logger.StreamOutput(stdoutPipe, onStdout, wg, stdoutChannel)
+	go logger.StreamOutput(stderrPipe, onStderr, wg, stderrChannel)
 
 	wg.Wait()
 	err = cmd.Wait()
-
-	// Optionally dump the output in the event of an error
-	if outputChan != nil {
-		close(outputChan)
-	}
-	if err != nil && printOutputOnError {
-		logger.Log.Errorf("Call to %s returned error, last %d lines of output:", cmd.Args, outputChanBufferSize)
-		for line := range outputChan {
-			logger.Log.Warn(line)
-		}
-	}
 
 	return
 }
